moved homePage to pages dir
