+++ src/router/router.js
@@ -1,7 +1,7 @@
 
 import { createRouter, createWebHistory } from 'vue-router';
 
-import HomePage from '@/components/HomePage.vue'
+import HomePage from '@/pages/HomePage.vue'
 
 const routes = [
     {'path': '/', component: HomePage},
